requirements/base.txt
Django==2.2.1
djangorestframework==3.9.4
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
certifi==2019.3.9 
chardet==3.0.4 
idna==2.8 
requests==2.22.0 
urllib3==1.25.3
MarkupSafe==1.1.1 
coreapi==2.3.3 
coreschema==0.0.4 
itypes==1.1.0 
jinja2==2.10.1 
pyyaml==5.1 
uritemplate==3.0.0
